Refactor type definitions and centralize interfaces
- Moved non-prop interfaces/types to `lib/interfaces`, creating individual files for each.
- Updated imports across the application to reference the new interface locations.
- Removed redundant type definitions from various components and files.
- Enhanced type hygiene by ensuring all shared types are consistently defined and imported.
- Added new interfaces for `AiRecommendation`, `ApplianceType`, `ArchiveEntry`, `CompleteRates`, `DashboardDevice`, `EnergyLogRow`, `LegacyEnergyLogRow`, `MeralcoFixedCharges`, `MeralcoRateComponents`, `ProfileRow`, `SyncRunStatus`, and others.
- Updated the `phases-and-roadmap-plan.instructions.md` to reflect the new shared interface location.
