+++ components/providers/SupabaseProvider.types.ts
@@ -1,10 +1,5 @@
-import type { Session, User } from "@supabase/supabase-js";
+import type { Session } from "@supabase/supabase-js";
 import type { ReactNode } from "react";
-
-export type SupabaseContextType = {
-  user: User | null;
-  session: Session | null;
-};
 
 export type SupabaseProviderProps = {
   children: ReactNode;
