+++ lib/interfaces/AiRecommendation.ts
@@ -1,0 +1,6 @@
+export type AiRecommendation = {
+  message: string;
+  estimated_monthly_kwh: number;
+  estimated_monthly_cost: number;
+  suggested_budget: number;
+};
