+++ lib/interfaces/ApplianceType.ts
@@ -1,0 +1,1 @@
+export type ApplianceType = "refrigerator" | "aircon" | "tv" | "other";
