+++ lib/interfaces/ArchiveEntry.ts
@@ -1,0 +1,6 @@
+export type ArchiveEntry = {
+  monthLabel: string;
+  monthDate: Date;
+  title: string;
+  pdfUrl: string;
+};
