+++ lib/interfaces/AuthFormContext.ts
@@ -1,0 +1,1 @@
+export type AuthFormContext = "login" | "register" | "reset";
