+++ lib/interfaces/CompleteRates.ts
@@ -1,0 +1,11 @@
+export type CompleteRates = {
+  generation: number;
+  transmission: number;
+  system_loss: number;
+  distribution: number;
+  universal_charges: number;
+  fit_all: number;
+  vat_rate: number;
+  metering_charge: number;
+  supply_charge: number;
+};
