+++ lib/interfaces/DashboardDeviceRow.ts
@@ -1,0 +1,8 @@
+export type DashboardDeviceRow = {
+  id: string;
+  device_name: string;
+  mac_address: string;
+  is_online: boolean | null;
+  appliance_type: string | null;
+  relay_state: boolean | null;
+};
