+++ lib/interfaces/DeviceDetailDeviceRow.ts
@@ -1,0 +1,6 @@
+export type DeviceDetailDeviceRow = {
+  id: string;
+  device_name: string;
+  mac_address: string;
+  relay_state: boolean | null;
+};
