+++ lib/interfaces/ExistingRateRow.ts
@@ -1,0 +1,12 @@
+export type ExistingRateRow = {
+  effective_month: string;
+  vat_rate: number;
+  generation: number;
+  transmission: number;
+  system_loss: number;
+  distribution: number;
+  universal_charges: number;
+  fit_all: number;
+  metering_charge: number;
+  supply_charge: number;
+};
