+++ lib/interfaces/InsightResult.ts
@@ -1,0 +1,4 @@
+export type InsightResult = {
+  message: string;
+  insight_type: string;
+};
