+++ lib/interfaces/InsightState.ts
@@ -1,0 +1,5 @@
+export type InsightState = {
+  loading: boolean;
+  message: string | null;
+  error: string | null;
+};
