+++ lib/interfaces/InsightType.ts
@@ -1,0 +1,5 @@
+export type InsightType =
+  | "budget_alert"
+  | "weekly_recap"
+  | "anomaly_alert"
+  | "cost_optimizer";
