+++ lib/interfaces/LatestReadingRow.ts
@@ -1,0 +1,8 @@
+export type LatestReadingRow = {
+  device_id: string;
+  average_watts: number | string | null;
+  voltage_v: number | string | null;
+  current_a: number | string | null;
+  energy_kwh: number | string | null;
+  recorded_at: string | null;
+};
