+++ lib/interfaces/LegacyEnergyLogRow.ts
@@ -1,0 +1,5 @@
+export type LegacyEnergyLogRow = {
+  energy_kwh: number | string;
+  average_watts: number | string | null;
+  recorded_at: string | null;
+};
