+++ lib/interfaces/MeralcoFixedCharges.ts
@@ -1,0 +1,4 @@
+export type MeralcoFixedCharges = {
+  meteringCharge: number;
+  supplyCharge: number;
+};
