+++ lib/interfaces/MeralcoRateComponents.ts
@@ -1,0 +1,8 @@
+export type MeralcoRateComponents = {
+  generation: number;
+  transmission: number;
+  systemLoss: number;
+  distribution: number;
+  universalCharges: number;
+  fitAll: number;
+};
