+++ lib/interfaces/MeralcoRatesRow.ts
