+++ lib/interfaces/MinMax.ts
@@ -1,0 +1,4 @@
+export type MinMax = {
+  min: number;
+  max: number;
+};
