+++ lib/interfaces/NumericLike.ts
@@ -1,0 +1,1 @@
+export type NumericLike = number | string | null | undefined;
