+++ lib/interfaces/ParsedRates.ts
@@ -1,0 +1,11 @@
+export type ParsedRates = {
+  generation: number | null;
+  transmission: number | null;
+  system_loss: number | null;
+  distribution: number | null;
+  universal_charges: number | null;
+  fit_all: number | null;
+  vat_rate: number | null;
+  metering_charge: number | null;
+  supply_charge: number | null;
+};
