+++ lib/interfaces/ProfileRow.ts
@@ -1,0 +1,3 @@
+export type ProfileRow = {
+  monthly_budget_php: number | string | null;
+};
