+++ lib/interfaces/SupabaseContextType.ts
@@ -1,0 +1,6 @@
+import type { Session, User } from "@supabase/supabase-js";
+
+export type SupabaseContextType = {
+  user: User | null;
+  session: Session | null;
+};
