+++ lib/interfaces/SyncRunStatus.ts
@@ -1,0 +1,1 @@
+export type SyncRunStatus = "success" | "failed" | "dry_run";
