+++ lib/interfaces/TimedReading.ts
@@ -1,0 +1,5 @@
+export type TimedReading = {
+  energyKwh: number;
+  timestampMs: number;
+  dayKey: string;
+};
