+++ lib/interfaces/UsageByDeviceDayRow.ts
@@ -1,0 +1,5 @@
+export type UsageByDeviceDayRow = {
+  device_id: string;
+  day_key: string;
+  usage_kwh: number | string;
+};
