+++ lib/interfaces/UsageByDeviceRow.ts
@@ -1,0 +1,4 @@
+export type UsageByDeviceRow = {
+  device_id: string;
+  usage_kwh: number | string;
+};
