+++ lib/interfaces/WeeklyUsagePoint.ts
@@ -1,0 +1,5 @@
+export type WeeklyUsagePoint = {
+  day: string;
+  thisWk: number;
+  lastWk: number;
+};
